py_sample/sample_load_log_config/pyproject.toml
[tool.poetry]
name = "sample_load_log_config"
version = "0.1.0"
description = ""
authors = ["wararaki <ky7.ott.w@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
PyYaml = "^5.3"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
